refactor: update imports and move functions out of MCPToolsComponent (#8976)
